+++ packages/react-data-grid-react-window/src/components/DataGrid/index.ts
@@ -1,2 +1,4 @@
 export * from './DataGrid';
+export * from './DataGrid.types';
+export * from './renderDataGrid';
 export * from './useDataGrid';
