+++ packages/react-data-grid-react-window/src/components/DataGridBody/index.ts
@@ -2,3 +2,4 @@ export * from './DataGridBody';
 export * from './DataGridBody.types';
 export * from './renderDataGridBody';
 export * from './useDataGridBody';
+export * from './useDataGridBodyStyles.styles';
