+++ packages/react-data-grid-react-window/src/components/DataGridHeader/index.ts
@@ -1,2 +1,3 @@
 export * from './DataGridHeader';
 export * from './useDataGridHeader';
+export * from './useDataGridHeaderStyles.styles';
