+++ packages/react-data-grid-react-window/src/components/DataGridRow/index.ts
@@ -1,2 +1,3 @@
 export * from './DataGridRow';
+export * from './useDataGridRow';
 export * from './useDataGridRow.styles';
